restructure setup component
