+++ pika/src/App.css
@@ -1,0 +1,17 @@
+.content-wrapper{
+	width: 95%;
+	margin: 20px auto;
+}
+
+.content-wrapper h1{
+	font-size: 50px;
+	color: #343a40;
+	text-align: center;
+	font-family: 'Source Sans Pro', sans-serif;
+}
+
+.content-wrapper h2{
+	font-weight: bold;
+	color: #646464;
+	font-family: 'Roboto', sans-serif;
+}
