+++ pika/src/Components/SetupComponents/SearchNodesSection.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+
+class SearchNodesSection extends Component {
+
+	render() {
+		return (
+			<div>
+				<h2>Search for Nodes</h2>
+			</div>
+		);
+	}
+	
+}
+
+export default SearchNodesSection;
