+++ pika/src/Components/SetupComponents/Setup.js
@@ -1,11 +1,16 @@
 import React, { Component } from 'react';
 
+import SearchNodesSection from './SearchNodesSection';
+
+import './css/setup.css';
+
 class Setup extends Component {
 
 	render() {
 		return (
 			<div>
 				<h1>Setup</h1>
+				<SearchNodesSection />
 			</div>
 		);
 	}
